Add movie detail fragment and layout
