+++ app/src/main/java/com/enginebai/moviehunt/ui/list/MovieListFragment.kt
@@ -13,7 +13,7 @@ import com.enginebai.base.view.BaseFragment
 import com.enginebai.moviehunt.R
 import com.enginebai.moviehunt.data.local.MovieModel
 import com.enginebai.moviehunt.ui.MovieClickListener
-import com.enginebai.moviehunt.ui.detail.MovieDetailFragment
+import com.enginebai.moviehunt.ui.detail.MovieDetailFragmentV1
 import com.enginebai.moviehunt.utils.openFragment
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
@@ -40,7 +40,7 @@ class MovieListFragment : BaseFragment(), MovieClickListener {
     }
 
     override fun onMovieClicked(movieId: String) {
-        activity?.openFragment(MovieDetailFragment.newInstance(movieId), true)
+        activity?.openFragment(MovieDetailFragmentV1.newInstance(movieId), true)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
