+++ app/src/main/res/layout/fragment_movie_detail.xml
@@ -1,112 +1,46 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    android:fitsSystemWindows="true">
 
-    <data>
-        <variable
-            name="viewModel"
-            type="com.enginebai.moviehunt.ui.detail.MovieDetailViewModel" />
-    </data>
-    <ScrollView
-        android:background="@color/background"
-        android:clickable="true"
-        android:focusable="true"
+    <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content">
 
-        <androidx.constraintlayout.widget.ConstraintLayout
+        <com.google.android.material.appbar.CollapsingToolbarLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-
-            <ImageView
-                android:id="@+id/imagePoster"
+            android:layout_height="match_parent"
+            >
+            <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:scaleType="centerCrop"
-                android:src="@color/darkBlue"
-                app:imageUrl="@{viewModel.posterUrl}"
-                app:error="@{@color/darkBlue}"
-                app:placeholder="@{@color/darkBlue}"
-                app:layout_constraintDimensionRatio="3:4"
-                app:layout_constraintTop_toTopOf="parent" />
+                android:layout_height="wrap_content">
+                <ImageView
+                    android:layout_width="0dp"
+                    android:layout_height="0dp"
+                    android:id="@+id/imagePoster"
+                    app:layout_constraintTop_toTopOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintDimensionRatio="3:4"
+                    android:src="@color/cardBackground"
+                    />
 
-            <ImageButton
-                android:id="@+id/buttonBack"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_margin="@dimen/size_16"
-                android:background="?selectableItemBackground"
-                android:src="@drawable/arrow_left"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                />
+                <androidx.appcompat.widget.Toolbar
+                    android:layout_width="match_parent"
+                    android:layout_height="?attr/actionBarSize"
+                    app:layout_constraintTop_toTopOf="parent"
+                    />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+        </com.google.android.material.appbar.CollapsingToolbarLayout>
 
-            <TextView
-                android:id="@+id/textTitle"
-                style="@style/TitleText.Large"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@{viewModel.title, default=@string/placeholder}"
-                android:padding="@dimen/size_16"
-                app:layout_constraintTop_toBottomOf="@+id/imagePoster"
-                tools:text="Star Wars: IV" />
+    </com.google.android.material.appbar.AppBarLayout>
 
-            <RatingBar
-                android:id="@+id/ratingBar"
-                style="@style/RatingBar"
-                android:rating="@{viewModel.rating}"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="@dimen/size_16"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/textTitle" />
+    <com.airbnb.epoxy.EpoxyRecyclerView
+        android:id="@+id/listContent"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"/>
 
-            <TextView
-                android:id="@+id/textVoteCount"
-                style="@style/VoteCountText"
-                android:text="@{@string/vote_count(viewModel.voteCount), default=@string/placeholder}"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                app:layout_constraintBottom_toBottomOf="@+id/ratingBar"
-                app:layout_constraintStart_toEndOf="@+id/ratingBar"
-                app:layout_constraintTop_toTopOf="@+id/ratingBar"
-                tools:text="2.6k reviews" />
-
-            <TextView
-                android:id="@+id/textDuration"
-                style="@style/ContextText"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@{viewModel.duration, default=@string/placeholder}"
-                android:drawableStart="@drawable/clock"
-                android:paddingStart="@dimen/size_16"
-                android:paddingEnd="@dimen/size_16"
-                app:layout_constraintTop_toBottomOf="@+id/textVoteCount"
-                tools:text="2 h 14 m" />
-
-            <TextView
-                android:id="@+id/textReleaseDate"
-                style="@style/ContextText"
-                android:text="@{viewModel.releaseDate, default=@string/placeholder}"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:drawableStart="@drawable/calendar"
-                android:paddingStart="@dimen/size_16"
-                android:paddingEnd="@dimen/size_16"
-                app:layout_constraintTop_toBottomOf="@+id/textDuration"
-                tools:text="2019/12/31" />
-
-            <TextView
-                android:id="@+id/textOverview"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@{viewModel.overview, default=@string/placeholder}"
-                android:padding="@dimen/size_16"
-                app:layout_constraintTop_toBottomOf="@+id/textReleaseDate"
-                tools:text="When it's discovered that the evil Emperor Palpatine did not die at the hands of Darth Vader, the rebels must race against the clock to find out his whereabouts. Finn and Poe lead the Resistance to put a stop to the First Order's plans to form a new Empire, while Rey anticipates her inevitable confrontation with Kylo Ren." />
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </ScrollView>
-</layout>
+</androidx.coordinatorlayout.widget.CoordinatorLayout>
